Ajustes en AuditorEditView
